insert con foto y datos hecho

+++ Datos/Users.cs
@@ -13,12 +13,12 @@ namespace Datos
         public string Nombre { get; set; }
         public string Apellido { get; set; }
         public decimal Altura { get; set; }
-        public int Documento { get; set; }
+        public long Documento { get; set; }
         public string FotoDePerfil { get; set; }
         public DateTime FechaDeNacimiento { get; set; }
         public bool Pago { get; set; }
         public int Peso { get; set; }
-        public int Contacto { get; set; }
+        public long Contacto { get; set; }
         public string IdCompra { get; set; }
